requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
image==1.5.33
pillow==10.0.0
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
